+++ changelog.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
 ## [0.2.4] [Github repo]
+### Added
+- `BLE_DEVICE` for bluetooth low energy devices
 ## [0.2.3] - 2020-05-25
 ### Fix
 - Serial device `is_reachable` method fix